feat: add content-type filtering utility
